refactor(frontend): split OrdererWorkflowSection and PartnerSection into sub-components
OrdererWorkflowSection (1,371 -> 306 lines parent + 6 sub-components):
- MemoCard: reusable memo card with edit/delete
- CallLogTab: call log with talk script accordion and memo list
- EvaluationTab: evaluation records with sort/add/edit
- TranscriptionTab: audio transcription display
- DocumentsTab: email templates and file upload
- Shared types and style constants

PartnerSection (1,357 -> 138 lines parent + 4 sub-components):
- PartnerCard: expandable partner card with tabs
- PartnerFilters: StatusChip, SurveyApprovedButton, FilterChips
- SentDocumentsSection: sent documents with assignee
- templates: script templates and placeholder replacement

Pure structural refactor - no behavior changes. TypeScript strict mode
compilation verified with zero errors.
